Fix toast adds scrollbar when appearing
Jquery slide adds the toast outside of the viewport BUT body
expands horizontally to accomodate the presence of the toast
which causes a scrollbar to appear. Can't make body overflow hidden
otherwise other legitimate objects won't be able to expand the view.

So we animate the width of the toast instead and make the white space
no wrap so it appears to be sliding in when in fact it's the width
that grows.